pyproject.toml
[project]
name = "retreiverqa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=26.3.1",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
]
